Add more comments, move test class
