+++ flow-polymer-template/src/test/java/com/vaadin/flow/component/polymertemplate/HasCurrentService.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component;
+package com.vaadin.flow.component.polymertemplate;
 
 import static org.junit.Assert.assertNull;
