+++ flow-polymer-template/src/test/java/com/vaadin/flow/templatemodel/TemplateModelProxyHandlerTest.java
@@ -4,8 +4,8 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.mockito.Mockito;
 
-import com.vaadin.flow.component.HasCurrentService;
 import com.vaadin.flow.component.UI;
+import com.vaadin.flow.component.polymertemplate.HasCurrentService;
 import com.vaadin.flow.function.DeploymentConfiguration;
 import com.vaadin.flow.server.VaadinService;
 import com.vaadin.flow.templatemodel.TemplateModelTest.EmptyModel;
