+++ flow-polymer-template/src/test/java/com/vaadin/flow/templatemodel/TemplateModelWithEncodersTest.java
@@ -13,7 +13,7 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.mockito.Mockito;
 
-import com.vaadin.flow.component.HasCurrentService;
+import com.vaadin.flow.component.polymertemplate.HasCurrentService;
 import com.vaadin.flow.function.DeploymentConfiguration;
 import com.vaadin.flow.internal.StateNode;
 import com.vaadin.flow.internal.nodefeature.ElementPropertyMap;
